Proto-Type Complete
Auto Generated Blogs and Topics based on Business Type and Similar Topic.

Next Step will be to create a Web-Application or GUI for the App, but this is fully functional as is.

client Data is stored in client_data.csv, Blog topics are store in <mm-dd-yyyy>_blogTopics.csv and are used to generate Blogs for each business on the list and organized within their respective folder.